move now_or_never function to bevy_utils::futures
